requirements.txt
fastapi==0.115.6
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
numpy>=1.24.0
PyJWT>=2.8.0
google-auth>=2.25.0
requests>=2.31.0
sentry-sdk[fastapi]>=1.40.0
stripe>=7.0.0
